// src/common/types/enums/ModalType.ts
export enum ModalType {
  NONE = "",
  NEW_CAREER = "NEW_CAREER",
  EDIT_TEAMS = "EDIT_TEAMS",
  DELETE_CONFIRM = "DELETE_CONFIRM",
  ADD_CLUB_IMG = "ADD_CLUB_IMG",
  EDIT_CLUB = "EDIT_CLUB",
  SLIDE_UP_PANEL = "SLIDE_UP_PANEL",
  DELETE_PLAYER_CONFIRM = "DELETE_PLAYER_CONFIRM",
  SELL_PLAYER = "SELL_PLAYER",
  LOAN_PLAYER = "LOAN_PLAYER",
  RETURN_LOAN_CONFIRM = "RETURN_LOAN_CONFIRM",
  SEASON_CONFIGS = "SEASON_CONFIGS",
  ADD_BADGE_CLUB = "ADD_BADGE_CLUB",
}